+++ src/AvTranscoder/transcoder/Transcoder.hpp
@@ -58,45 +58,45 @@ public:
 	 * If offset is positive, the transcoder will generate black images or silence (depending on the type of stream) before the stream to process.
 	 * If offset is negative, the transcoder will seek in the stream and start process at this specific time.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const std::string& profileName = "", const double offset = 0 );
+	void add( const std::string& filename, const size_t streamIndex, const std::string& profileName = "", const float offset = 0 );
 	/*
 	 * @note If filename is empty, add a generated stream.
 	 * @note If filename is empty, profileName can't be empty (no sens to rewrap a generated stream).
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const std::string& profileName, ICodec& codec, const double offset = 0 );
+	void add( const std::string& filename, const size_t streamIndex, const std::string& profileName, ICodec& codec, const float offset = 0 );
 
 	/**
 	 * @brief Add a stream and set a custom profile
 	 * @note Profile will be updated, be sure to pass unique profile name.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const ProfileLoader::Profile& profile, const double offset = 0 );
+	void add( const std::string& filename, const size_t streamIndex, const ProfileLoader::Profile& profile, const float offset = 0 );
 	/*
 	 * @note If filename is empty, add a generated stream.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const ProfileLoader::Profile& profile, ICodec& codec, const double offset = 0  );
+	void add( const std::string& filename, const size_t streamIndex, const ProfileLoader::Profile& profile, ICodec& codec, const float offset = 0  );
 	
 	/**
 	 * @brief Add a stream and set a profile
 	 * @note If profileName is empty, rewrap.
 	 * @note If subStreamIndex is negative, no substream is selected it's the stream.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const std::string& profileName = "", const double offset = 0 );
+	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const std::string& profileName = "", const float offset = 0 );
 	/**
 	 * @note If filename is empty, add a generated stream.
 	 * @note If filename is empty, profileName can't be empty (no sens to rewrap a generated stream).
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const std::string& profileName, ICodec& codec, const double offset = 0  );
+	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const std::string& profileName, ICodec& codec, const float offset = 0  );
 
 	/**
 	 * @brief Add a stream and set a custom profile
 	 * @note Profile will be updated, be sure to pass unique profile name.
 	 * @note If subStreamIndex is negative, no substream is selected it's the stream.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, const double offset = 0 );
+	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, const float offset = 0 );
 	/**
 	 * @note If filename is empty, add a generated stream.
 	 */
-	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, ICodec& codec, const double offset = 0  );
+	void add( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, ICodec& codec, const float offset = 0  );
 
 	/**
 	 * @brief Add the stream
@@ -152,14 +152,14 @@ public:
 	void setProcessMethod( const EProcessMethod eProcessMethod, const size_t indexBasedStream = 0, const double outputDuration = 0 );
 
 private:
-	void addRewrapStream( const std::string& filename, const size_t streamIndex, const double offset );
+	void addRewrapStream( const std::string& filename, const size_t streamIndex, const float offset );
 
-	void addTranscodeStream( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const double offset );
-	void addTranscodeStream( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, const double offset = 0 );
+	void addTranscodeStream( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const float offset );
+	void addTranscodeStream( const std::string& filename, const size_t streamIndex, const int subStreamIndex, const ProfileLoader::Profile& profile, const float offset = 0 );
 
 	void addDummyStream( const ProfileLoader::Profile& profile, const ICodec& codec );
 
-	InputFile* addInputFile( const std::string& filename, const size_t streamIndex, const double offset );
+	InputFile* addInputFile( const std::string& filename, const size_t streamIndex, const float offset );
 
 	ProfileLoader::Profile getProfileFromFile( InputFile& inputFile, const size_t streamIndex );  ///< The function analyses the inputFile
